# Improve performance of shroud updating.