Fixes #17377 
Fixes #16393

This improves the performance of shroud updating by only updating  each cell at most once per tick and getting rid of as much redundant memory allocations as possible. Especially operations like `HashSet.UnionWith` appear to be a major bottleneck. I tried to replace dynamic data structures with fixed-size data structures where reasonable (`CellLayer` in particular). I tried to reduce the number of operations in general by eliminating redundancies between `FrozenActorLayer`, `ShroudRenderer,` `RadarWidget` and `PlayerRadarTerrain`.

As a testcase I tried building 60 yaks on Pool Party: 
- On bleed framerate on my machine drops to 20 fps (down from 60 fps frame limited).
- If I set  `MoveRecalculationThreshold` to 0 (effectively reverting #16953) I get ~45 fps.
- On this PR I get 57&ndash;60 fps.