@pchote 2019-11-25
A lot is going on here, can you please give a rough overview for each of the key changes here? (e.g. removing the bins from FrozenActorLayer, the touched/invalidation changes in Shroud, the changes in `RenderShroud`/`PlayerRadarTerrain`, etc).

I guess we will need a second PR targeting `prep-1908` specifically, thanks to #17076.