@teinarss 2019-11-25
One regression from #17076 is that the Radar dont update the same way when you Spectate as the player. 
![image](https://user-images.githubusercontent.com/19173272/69578742-65e04700-0fd1-11ea-8fd3-6bae8740cd40.png)