@reaperrr 2019-11-25
> Or are we happy to go with the much simpler MoveRecalculationThreshold: 0 bandaid for the hotfix?

Just my 2 cents, but I'd prefer that we should go with the bandaid for the hotfix. Our track-record of finding all bugs/regressions in prs of this size hasn't been that great, and having to adapt and review another PR specifically for prep likely just increases the chance of something slipping through into that release, regardless of playtests.