@tovl 2019-11-25
Key changes:
- `AddSource`/`RemoveSource`/etc. no longer trigger an Invalidation.
- Invalidation is moved into `Shroud.Tick` which loops over all map cells once per tick.
- `Touched` is used to track which cells have been affected in any way since the last tick. If a cell is not touched the invalidation for that cell is skipped immediately.
- `oldResolvedType` is used to keep track of the shroud state of the previous tick.
- `CellsChanged` has been removed. The `CellsChanged` actions would previously make `ShroudRenderer`, `FrozenActorLayer`, `RadarWidget` and `PlayerRadarTerrain` update their own internal dirty cells cache. They would then act on those cells in their own tick. The delayed action was needed when invalidation happened multiple times per tick but all the updating of redundant sets has major performance costs of its own.
- `OnShroudChanged` has been added instead. It directly triggers the actual updates done by those classes on a specific cell.
- When the evaluated shroud state is the same as the previous tick for a cell, the `OnShroudChanged` actions are skipped for that cell, avoiding redundant work.
- Code in those classes that cycles through cells/bins independently of `Shroud` and evaluates whether a given cell needs and update and then updates that cell have mostly been removed. All updates are directly triggered by `Shroud.Tick` instead.
- `ShroudRenderer` still needs to cycle through cells an additional time before rendering. This is because all surrounding cells of a changed cell need to be updated as well to get the right sprites. The `OnShroudChanged` action here is used to mark all surrounding cells as dirty in a separate `CellLayer`.
- The actions in `OnShroudChanged` that need to happen when the renderplayer is changed where awkwardly mixed in with the actions that need to happen every tick. That has been disentangled and moved to a `world.RenderPlayerChanged` action following the example set by `RadarWidget`.