@tovl 2019-11-25
This is a complicated restructuring with a high chance of regressions missed during review so I wouldn't recommend it for a quick fix. The bandaid fix will have to do if we go the route of a bugfix release.