@pchote 2019-11-26
Significant chunks (possibly most of) `touched` is going to remain false each tick. Wouldn't it make sense to keep a coarser `bins` array to track these regions at a coarser level so we don't need to enumerate the whole layer each tick?