@tovl 2019-11-26
It is difficult to say without trying it out. It might squeeze a few more frames out put it might also regress the performance with a larger amount of units because of the extra operations we incur in `AddSource`/`RemoveSource`. I should note that `touched` itself is only responsible for a relatively small (but still significant) part of the performance gain here. Even when we do the invalidation for every cell on every tick, we still get most of the performance gains as long as we still skip the `OnShroudChanged` actions.

I also tried doubling the amount of yaks in the testcase to 120 and moving them as a single flock. That dropped the framerate to about 40&ndash;45 (10 on bleed). Moving them as a single flock means that the number of cells touched is mostly the same as with 60 yaks which means it is the per-unit operations that become the dominant bottleneck with these numbers. So I think the biggest performance gains at this point can be had in the projection part or the  `AddSource`/`RemoveSource` part.