@pchote 2019-11-26
Ok. We can potentially halve the per-unit operations in RA by reworking the gap generator implementation to not need two `RevealsShroud` - something I plan to look into for a future PR.