@tovl 2019-11-26
> One regression from #17076 is that the Radar dont update the same way when you Spectate as the player.

Fixed.