@RoosterDragon 2019-11-27
This makes sense as a high-level change for centralizing shroud changes into one pass and all the changes you listed flow naturally from that - the rest would just be nitpicking over low-level specifics.

My main question at a high-level would be if delaying the updates affects any Tick logic in an adverse way. Currently if a unit moves into a new cell, it reveals the shroud immediately, whereas now this would be delayed by a tick so other decision making would see different outcomes when querying the visibility states of the cell within that tick.

That's the major blocker as to whether this is a valid optimization to apply - if we can satisfy ourselves that the delays doesn't have any unfortunate knock over behaviour we'd be good.

A theoretical example of what I mean (I haven't checked if something like this would actually apply)
- A unit with identical sight and attack range moves into range of an enemy.
- Under previous behaviour, the enemy is revealed so would be able to be attacked.
- Under new behaviour, the shroud is still hidden so the enemy can't be seen.
- Unit continues moving.
- Next tick, the enemy becomes visible, but our unit continues moving into the next cell over several ticks.
- End result: Units end up moving one cell closer than previously, which might seem odd or affect balance.